+++ src/main/resources/db/migration/V20250718.000008__video_provider.sql
@@ -1,7 +1,7 @@
 -- Adds integer (ordinal) columns for provider to videos table.
 
 ALTER TABLE videos
-    ADD COLUMN IF NOT EXISTS provider INTEGER NOT NULL DEFAULT 0;
+    ADD COLUMN IF NOT EXISTS provider SMALLINT NOT NULL DEFAULT 0;
 
 -- add simple check constraints to bound future values.
 -- Adjust upper bounds when you add more enum constants.
